requirements.txt
peewee
aiogram
aiohttp>=3.9.0
python-dotenv
openai>=1.40.0
httpx>=0.27.0
